Major refactoring of the localftp index fetcher
The localftp index fetcher now uses only utility functions and classes to manage access to the index and searches, with cache or not.
Introduces an ``indexstore`` that should be able to manage local and online searches on the Argo index.
Searches are based on a new index_filter_wmo that inherits from a index_filter_proto. These works in memory and will be reusable for online and offline indexstores.
