+++ argopy/utilities.py
@@ -13,6 +13,7 @@ import requests
 import urllib.request
 import io
 import json
+import numpy as np
 import xarray as xr
 import pandas as pd
 from IPython.core.display import display, HTML
@@ -31,13 +32,327 @@ path2pkl = pkg_resources.resource_filename('argopy', 'assets/')
 from argopy.errors import ErddapServerError, FileSystemHasNoCache, CacheFileNotFound
 from argopy.options import OPTIONS
 
+from abc import ABC, abstractmethod
+import hashlib
+
+
+class index_filter_proto(ABC):
+    def __init__(self):
+        pass
+
+    @abstractmethod
+    def run(self):
+        """ Take a _io.TextIOWrapper and return filter results as string
+
+        Parameters
+        ----------
+        index_file: _io.TextIOWrapper
+
+        Returns
+        -------
+        csv rows matching the request, as a in-memory string. Or None.
+        """
+        pass
+
+    @abstractmethod
+    def uri(self):
+        """ Return a name for one specific filter run """
+        pass
+
+    @property
+    def sha(self):
+        return hashlib.sha256(self.uri().encode()).hexdigest()
+
+
+class index_filter_wmo(index_filter_proto):
+    def __init__(self, WMO: list = [], CYC=None, **kwargs):
+        """ Create Argo index filter for WMOs/CYCs
+
+            Parameters
+            ----------
+            WMO : list(int)
+                The list of WMOs to search
+            CYC : int, np.array(int), list(int)
+                The cycle numbers to search for each WMO
+        """
+        if isinstance(WMO, int):
+            WMO = [WMO]  # Make sure we deal with a list
+        if isinstance(CYC, int):
+            CYC = np.array((CYC,), dtype='int')  # Make sure we deal with an array of integers
+        if isinstance(CYC, list):
+            CYC = np.array(CYC, dtype='int')  # Make sure we deal with an array of integers
+        self.WMO = sorted(WMO)
+        self.CYC = CYC
+
+    def uri(self):
+        if len(self.WMO) > 1:
+            listname = ["WMO%i" % i for i in sorted(self.WMO)]
+            if isinstance(self.CYC, (np.ndarray)):
+                [listname.append("CYC%0.4d" % i) for i in sorted(self.CYC)]
+            listname = "_".join(listname)
+        elif len(self.WMO) == 0:
+            if isinstance(self.CYC, (np.ndarray)):
+                listname = ["AllWMOs"]
+                [listname.append("CYC%0.4d" % i) for i in sorted(self.CYC)]
+            else:
+                listname = ["FULL"]
+            listname = "_".join(listname)
+        else:
+            listname = "WMO%i" % self.WMO[0]
+            if isinstance(self.CYC, (np.ndarray)):
+                listname = [listname]
+                [listname.append("CYC%0.4d" % i) for i in sorted(self.CYC)]
+                listname = "_".join(listname)
+        if len(listname) > 256:
+            listname = hashlib.sha256(listname.encode()).hexdigest()
+        return listname
+
+    def run(self, index_file):
+        """ Run search on an Argo index file
+
+        Parameters
+        ----------
+        index_file: _io.TextIOWrapper
+
+        Returns
+        -------
+        csv rows matching the request, as a in-memory string. Or None.
+        """
+
+        def search_one_wmo(index, wmo):
+            """ Search for a WMO in the argo index file
+
+            Parameters
+            ----------
+            index_file: _io.TextIOWrapper
+            wmo: int
+
+            Returns
+            -------
+            csv chunk matching the request, as a string. Or None
+            """
+            index.seek(0)
+            results = ""
+            il_read, il_loaded, il_this = 0, 0, 0
+            for line in index:
+                il_this = il_loaded
+                # if re.search("/%i/" % wmo, line.split(',')[0]):
+                if "/%i/" % wmo in line:  # much faster than re
+                    # Search for the wmo at the beginning of the file name under: /<dac>/<wmo>/profiles/
+                    results += line
+                    il_loaded += 1
+                if il_this == il_loaded and il_this > 0:
+                    break  # Since the index is sorted, once we found the float, we can stop reading the index !
+                il_read += 1
+            if il_loaded > 0:
+                return results
+            else:
+                return None
+
+        def search_any_wmo_cyc(index, cyc):
+            """ Search for a WMO in the argo index file
+
+            Parameters
+            ----------
+            index_file: _io.TextIOWrapper
+            cyc: array of integers
+
+            Returns
+            -------
+            csv chunk matching the request, as a string. Or None
+            """
+
+            def search_this(this_line):
+                # return np.any([re.search("%0.3d.nc" % c, this_line.split(',')[0]) for c in cyc])
+                return np.any(["%0.3d.nc" % c in this_line for c in cyc])
+                if np.all(cyc >= 1000):
+                    def search_this(this_line):
+                        # return np.any([re.search("%0.4d.nc" % c, this_line.split(',')[0]) for c in cyc])
+                        return np.any(["%0.4d.nc" % c in this_line for c in cyc])
+
+            index.seek(0)
+            results = ""
+            il_read, il_loaded, il_this = 0, 0, 0
+            for line in index:
+                il_this = il_loaded
+                if search_this(line):
+                    results += line
+                    il_loaded += 1
+                il_read += 1
+            if il_loaded > 0:
+                return results
+            else:
+                return None
+
+        def search_one_wmo_cyc(index, wmo, cyc):
+            """ Search for a WMO and CYC in the argo index file
+
+            Parameters
+            ----------
+            index: _io.TextIOWrapper
+            wmo: int
+            cyc: array of integers
+
+            Returns
+            -------
+            csv chunk matching the request, as a string. Or None
+            """
+            index.seek(0)
+            results = ""
+
+            # Look for the float:
+            il_read, il_loaded, il_this = 0, 0, 0
+            for line in index:
+                il_this = il_loaded
+                # if re.search("/%i/" % wmo, line.split(',')[0]):
+                if "/%i/" % wmo in line:  # much faster than re
+                    results += line
+                    il_loaded += 1
+                if il_this == il_loaded and il_this > 0:
+                    break  # Since the index is sorted, once we found the float, we can stop reading the index !
+                il_read += 1
+
+            # Then look for the profile:
+            if results:
+                def search_this(this_line):
+                    # return np.any([re.search("%0.3d.nc" % c, this_line.split(',')[0]) for c in cyc])
+                    return np.any(["%0.3d.nc" % c in this_line for c in cyc])
+                if np.all(cyc >= 1000):
+                    def search_this(this_line):
+                        # return np.any([re.search("%0.4d.nc" % c, this_line.split(',')[0]) for c in cyc])
+                        return np.any(["%0.4d.nc" % c in this_line for c in cyc])
+                il_loaded, cyc_results = 0, ""
+                for line in results.split():
+                    if search_this(line):
+                        il_loaded += 1
+                        cyc_results += line + "\n"
+            if il_loaded > 0:
+                return cyc_results
+            else:
+                return None
+
+        def full_load(index):
+            """ Return the full argo index file (without header)
+
+            Parameters
+            ----------
+            index: _io.TextIOWrapper
+
+            Returns
+            -------
+            csv index, as a string
+            """
+            index.seek(0)
+            results = ""
+            for line in index:
+                if line[0] != '#':
+                    break
+            return index.read()
+
+        if len(self.WMO) > 1:
+            if isinstance(self.CYC, (np.ndarray)):
+                return "".join([r for r in [search_one_wmo_cyc(index_file, w, self.CYC) for w in self.WMO] if r])
+            else:
+                return "".join([r for r in [search_one_wmo(index_file, w) for w in self.WMO] if r])
+        elif len(self.WMO) == 0:  # Search for cycle numbers only
+            if isinstance(self.CYC, (np.ndarray)):
+                return search_any_wmo_cyc(index_file, self.CYC)
+            else:
+                # No wmo, No cyc, return the full index:
+                return full_load(index_file)
+        else:
+            if isinstance(self.CYC, (np.ndarray)):
+                return search_one_wmo_cyc(index_file, self.WMO[0], self.CYC)
+            else:
+                return search_one_wmo(index_file, self.WMO[0])
+
+
+class indexstore():
+    """" Use to manage access to a local Argo index and searches """
+    def __init__(self,
+                 cache: bool = False,
+                 cachedir: str = "",
+                 index_file: str = "ar_index_global_prof.txt",
+                 **kw):
+        """ Create a file storage system for local file requests
+
+            Parameters
+            ----------
+            cache : bool (False)
+            cachedir : str (from OPTIONS)
+
+        """
+        self.index_file = index_file
+        self.cache = cache
+        self.cachedir = OPTIONS['cachedir'] if cachedir == '' else cachedir
+        self.fs = {}
+        self.fs['index'] = filestore(cache, cachedir)
+        if not cache:
+            self.fs['search'] = fsspec.filesystem("memory")
+        else:
+            self.fs['search'] = fsspec.filesystem("filecache",
+                                                 target_protocol='memory',
+                                                 cache_storage=OPTIONS['cachedir'] if cachedir == '' else cachedir,
+                                                 expiry_time=86400,
+                                                 cache_check=10)
+            self.fs['search'].load_cache()
+
+    def in_cache(self, fs, uri):
+        """ Return true if uri is cached """
+        if not uri.startswith(fs.target_protocol):
+            store_path = fs.target_protocol + "://" + uri
+        else:
+            store_path = uri
+        fs.load_cache()
+        return store_path in fs.cached_files[-1]
+
+    def in_memory(self, fs, uri):
+        """ Return true if uri is in the memory store """
+        return uri in fs.store
+
+    def open_index(self):
+        return self.fs['index'].open(self.index_file, "r")
+
+    def res2dataframe(self, results):
+        """ Convert a csv like string into a DataFrame
+
+            If one columns has a missing value, the row is skipped
+        """
+        return pd.DataFrame([x.split(',') for x in results.split('\n') if ",," not in x],
+                            columns=['file', 'date', 'latitude', 'longitude', 'ocean', 'profiler_type',
+                                     'institution', 'date_update']) \
+                   .astype({'file': np.str,
+                            'date': np.datetime64,
+                            'latitude': np.float32,
+                            'longitude': np.float32,
+                            'ocean': np.str,
+                            'profiler_type': np.str,
+                            'institution': np.str,
+                            'date_update': np.datetime64})[:-1]
+
+    def open_dataframe(self, search_cls):
+        uri = search_cls.uri()
+        with self.open_index() as f:
+            if self.cache and (self.in_cache(self.fs['search'], uri) or self.in_memory(self.fs['search'], uri)):
+                # print('Search already in memory, loading:', uri)
+                with self.fs['search'].open(uri, "r") as of:
+                    df = self.res2dataframe(of.read())
+            else:
+                # print('Running search from scratch ...')
+                # Run search:
+                results = search_cls.run(f)
+                # and save results for caching:
+                if self.cache:
+                    with self.fs['search'].open(uri, "w") as of:
+                        of.write(results)  # This happens in memory
+                df = self.res2dataframe(results)
+        return df
+
 
 class filestore():
     """Wrapper around fsspec file stores
 
         https://filesystem-spec.readthedocs.io/en/latest/api.html#fsspec.implementations.local.LocalFileSystem
-
-        This wrapper is primarily used by the localftp data/index fetchers
 
     """
